Move Advanced Ai to separate files
